Allow context propagation through Lock operations
To enable functionality such as tracing http via context, we
need to make sure that any query or write operations are context-aware.

This commit adds a ctx field to LockOptions and Context()/WithContext()
methods in the same style as WriteOptions and QueryOptions. For each
write or query operation in the Lock lifecycle, we pass along the ctx
value to the WriteOptions or QueryOptions.